+++ static/audited_at.json
@@ -14,6 +14,7 @@
   "Primer::Alpha::Dropdown": "",
   "Primer::Alpha::Dropdown::Menu": "",
   "Primer::Alpha::Dropdown::Menu::Item": "",
+  "Primer::Alpha::HellipButton": "",
   "Primer::Alpha::HiddenTextExpander": "",
   "Primer::Alpha::Image": "",
   "Primer::Alpha::ImageCrop": "",
