+++ static/constants.json
@@ -216,6 +216,8 @@
       "summary"
     ]
   },
+  "Primer::Alpha::HellipButton": {
+  },
   "Primer::Alpha::HiddenTextExpander": {
   },
   "Primer::Alpha::Image": {
