+++ static/statuses.json
@@ -14,6 +14,7 @@
   "Primer::Alpha::Dropdown": "alpha",
   "Primer::Alpha::Dropdown::Menu": "alpha",
   "Primer::Alpha::Dropdown::Menu::Item": "alpha",
+  "Primer::Alpha::HellipButton": "alpha",
   "Primer::Alpha::HiddenTextExpander": "alpha",
   "Primer::Alpha::Image": "alpha",
   "Primer::Alpha::ImageCrop": "alpha",
@@ -77,7 +78,7 @@
   "Primer::Dropdown::Menu": "deprecated",
   "Primer::Dropdown::Menu::Item": "deprecated",
   "Primer::DropdownMenuComponent": "deprecated",
-  "Primer::HellipButton": "alpha",
+  "Primer::HellipButton": "deprecated",
   "Primer::IconButton": "deprecated",
   "Primer::LabelComponent": "deprecated",
   "Primer::LayoutComponent": "alpha",
